+++ offline-1/2105118/2105118_scopeTable.cpp
@@ -20,13 +20,13 @@ private:
     int sid ; // static variable to keep track of scope table id
     int bucketSize ;
     int collissionCount ;
-    unsigned long (*hashFunction)(string); // function pointer
+    unsigned long long (*hashFunction)(string); // function pointer
     symbolInfo** hashtable; // hashtable of symbolInfo
     ScopeTable* parentScope ; // pointer to parent scope, to be used as a stack
 
 public:
     
-    ScopeTable(int bucketSize, unsigned long (*hashFunction)( string))  {
+    ScopeTable(int bucketSize, unsigned long long (*hashFunction)( string))  {
         hashtable = new symbolInfo*[bucketSize];
         this->bucketSize = bucketSize;
         for (int i = 0; i < bucketSize; i++) {
@@ -91,7 +91,7 @@ public:
     
     symbolInfo * getLookUp ( string name, int &position){
         name = trim(name);
-        unsigned long index = getHash(name);
+        unsigned long long index = getHash(name);
         symbolInfo * hashPointer = hashtable[index];
         while ( hashPointer!= nullptr){
             if (hashPointer->getName() == name){
@@ -103,7 +103,7 @@ public:
         return nullptr;
     }
 
-    unsigned long getHash(string str){
+    unsigned long long getHash(string str){
         str = trim(str);
         return hashFunction(str) % bucketSize;
     }
@@ -115,7 +115,7 @@ public:
             return false;
         } 
         sInfo->setName(trim(sInfo->getName()));
-        unsigned long index = getHash(sInfo->getName());    
+        unsigned long long index = getHash(sInfo->getName());    
         int position = 1;
         int i = 1;
         symbolInfo* found = getLookUp(sInfo->getName(), i);
@@ -154,7 +154,7 @@ public:
     symbolInfo * LookUp(string name) {
         int position = 1;
         name = trim(name);
-        unsigned long index = getHash(name) +1 ;
+        unsigned long long index = getHash(name) +1 ;
         symbolInfo* hashPointer = getLookUp(name, position);
         if (hashPointer != nullptr) {
             fout << "\t\'" << name << "\' found in ScopeTable# " << sid << " at position " << index << ", " << position << endl;
@@ -171,7 +171,7 @@ public:
             return false;
         }
         name = trim(name);
-        unsigned long index = getHash(name);
+        unsigned long long index = getHash(name);
         if (getLookUp(name, i) == nullptr) {
              fout << "\tNot found in the current ScopeTable"<<endl;
             return false;
@@ -197,7 +197,7 @@ public:
         return false;
     }
 
-    unsigned long (*getHashFunction())(string) {
+    unsigned long long (*getHashFunction())(string) {
         return hashFunction;
     }
